Experiments with launcher
UI improvements for the launcher
- Increased the font and padding
- Added a new line on top of prompt to get better top padding

Removed gptel advice to kill. The default to open actual Gptel buffer is better.

One issue has persisted that I've failed to solve:
- Emacs window stealing focus when launcher is closed

This is most likely due to how macOS handles focus.
It seems to me that it tries to switchback to the application that created the popup.
Which in this case is emacs and it jumps back to the main window.

I've tried lot of things to solve this from varous frame params to using applescript.
But nothing has worked so far.

I also tried running it through daemon. It works fine but a new issue arises.
Things that require a emacs window such as searching through agenda,
 jumping to gptel buffer, etc are not possible as the daemon is running in the background.